Add tracking module
